backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai[google-genai,litellm]>=1.9.3",
    "fastapi>=0.129.0",
    "google-generativeai>=0.8.6",
    "pillow>=12.1.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
    "feedparser>=6.0.11",
    "apscheduler>=3.11.0",
    "pymongo>=4.12.0",
    "pytrends>=4.9.2",
    "python-telegram-bot>=21.0",
]
